+++ chainer/link.py
@@ -133,9 +133,6 @@ class Link(object):
             shape, dtype = _ensure_shape_dtype(value)
             self.add_param(name, shape, dtype=dtype)
 
-        self.grads_gathered = None
-        self.params_gathered = None
-
     @property
     def xp(self):
         """Array module for this link.
@@ -456,143 +453,6 @@ class Link(object):
         for link in self.links():
             for param in link._uninitialized_params.values():
                 param._zeroed = True
-
-    def size_num_grads(self):
-        """Count size of all gradient arrays
-
-        """
-        size = 0
-        num = 0
-        for param in self.params():
-            if param.size == 0:
-                continue
-            size += param.size
-            num += 1
-        return size, num
-
-    def _batch_memcpy(self):
-        return cuda.cupy.ElementwiseKernel(
-            'raw T ptrs, raw X info',
-            'raw float32 dst',
-            '''
-                int id_min = id_pre;
-                int id_max = num_src;
-                while (id_max - id_min > 1) {
-                    int id = (id_max + id_min) / 2;
-                    if (i < info[id]) id_max = id;
-                    else              id_min = id;
-                }
-                int id = id_min;
-                float *src = (float *)(ptrs[id]);
-                int i_dst = i;
-                int i_src = i;
-                if (id > 0) i_src -= info[id];
-                dst[i_dst] = 0;
-                if (src != NULL) {
-                    dst[i_dst] = src[i_src];
-                }
-                id_pre = id;
-            ''',
-            'batch_memcpy',
-            loop_prep='''
-                int num_src = info[0];
-                int id_pre = 0;
-            ''')
-
-    def gather_grads(self):
-        """Put together all gradient arrays and make a single array
-
-        Return:
-            cupy.ndarray
-        """
-        if self._cpu:
-            raise RuntimeError('Link.gather_grads works only on GPU.')
-
-        size, num = self.size_num_grads()
-        # print("size:{}, num:{}".format(size, num))
-
-        ptrs = numpy.empty(num, dtype=numpy.uint64)
-        info = numpy.empty(num + 1, dtype=numpy.int32)
-        info[0] = 0
-        i = 0
-        for param in self.params():
-            if param.size == 0:
-                continue
-            ptrs[i] = 0  # NULL pointer
-            if param._grad is not None:
-                ptrs[i] = param._grad.data.ptr
-            info[i + 1] = info[i] + param.size
-            i += 1
-        info[0] = num
-        # print("ptrs:\n{}".format(ptrs))
-        # print("info:\n{}".format(info))
-
-        ptrs = cuda.to_gpu(ptrs, stream=cuda.Stream.null)
-        info = cuda.to_gpu(info, stream=cuda.Stream.null)
-
-        batch_memcpy = self._batch_memcpy()
-
-        self.grads_gathered = batch_memcpy(ptrs, info, size=size)
-        return self.grads_gathered
-
-    def gather_params(self):
-        """Put together all gradient arrays and make a single array
-
-        Return:
-            cupy.ndarray
-        """
-        if self._cpu:
-            raise RuntimeError('Link.gather_params works only on GPU.')
-
-        size, num = self.size_num_grads()
-        # print("size:{}, num:{}".format(size, num))
-
-        ptrs = numpy.empty(num, dtype=numpy.uint64)
-        info = numpy.empty(num + 1, dtype=numpy.int32)
-        info[0] = 0
-        i = 0
-        for param in self.params():
-            if param.size == 0:
-                continue
-            ptrs[i] = 0  # NULL pointer
-            if param.data is not None:
-                ptrs[i] = param.data.data.ptr
-            info[i + 1] = info[i] + param.size
-            i += 1
-        info[0] = num
-        # print("ptrs:\n{}".format(ptrs))
-        # print("info:\n{}".format(info))
-
-        ptrs = cuda.to_gpu(ptrs, stream=cuda.Stream.null)
-        info = cuda.to_gpu(info, stream=cuda.Stream.null)
-
-        batch_memcpy = self._batch_memcpy()
-
-        self.params_gathered = batch_memcpy(ptrs, info, size=size)
-        return self.params_gathered
-
-    def scatter_grads(self, array):
-        """Put back contents of the specified array to the related gradient arrays
-
-        Args:
-            array (cupy.ndarray): gathered array created by gather_grads()
-        """
-        offset = 0
-        for param in self.params():
-            offset = param.scatter_grad(array, offset)
-
-    def scatter_params(self, array):
-        """Put back contents of the specified array to the related gradient arrays
-
-        Args:
-            array (cupy.ndarray): gathered array created by gather_grads()
-        """
-        offset = 0
-        for param in self.params():
-            offset = param.scatter_param(array, offset)
-
-    def cleargrads_gathered(self):
-        self.grads_gathered = None
 
     def addgrads(self, link):
         """Accumulates gradient values from given link.
